#61 Move InfinitySupport to its own file
